Import manually-reconstructed logo and head.
The logo was manually retraced from the 2385x1260 logo at
https://media.info/radio/stations/keith-community-radio/logo-archive.
Based on the artifacts/out-of-place edges, that image seems to be the
product of an Inkscape bitmap tracing. This and other misalignments lead
me to believe that the image was traced and never cleaned up, so I feel
pretty certain that I'm not acting against any intentional design
decisions in my tidying.